#10276 suppress filters/quickactions refreshes when data request comes from a ws

+++ frontend/src/actions/ViewActions.js
@@ -314,6 +314,7 @@ export function fetchDocument({
   pageLength,
   orderBy,
   isModal = false,
+  websocketRefresh = false,
 }) {
   return (dispatch, getState) => {
     dispatch(fetchDocumentPending(windowId, isModal));
@@ -327,8 +328,11 @@ export function fetchDocument({
     })
       .then((response) => {
         // remove the old filter from the store
-        const entityRelatedId = getEntityRelatedId({ windowId, viewId });
-        dispatch(deleteFilter(entityRelatedId));
+        if (!websocketRefresh) {
+          const entityRelatedId = getEntityRelatedId({ windowId, viewId });
+          dispatch(deleteFilter(entityRelatedId));
+        }
+
 
         dispatch(fetchDocumentSuccess(windowId, response.data, isModal));
 
@@ -345,26 +349,29 @@ export function fetchDocument({
 
         const state = getState();
         const view = getView(state, windowId, isModal);
-        const filterId = getEntityRelatedId({ windowId, viewId });
-        const activeFiltersCaptions = populateFiltersCaptions({
-          filterData: view.layout.filters,
-          filtersActive: response.data.filters,
-        });
-        const filtersActive = formatFilters({
-          filtersData: view.layout.filters,
-          filtersActive: response.data.filters,
-        });
 
-        dispatch(
-          createFilter({
-            filterId,
-            data: {
-              filterData: view.layout.filters, // set the proper layout for the filters
-              filtersActive,
-              activeFiltersCaptions,
-            },
-          })
-        );
+        if (!websocketRefresh) {
+          const filterId = getEntityRelatedId({ windowId, viewId });
+          const activeFiltersCaptions = populateFiltersCaptions({
+            filterData: view.layout.filters,
+            filtersActive: response.data.filters,
+          });
+          const filtersActive = formatFilters({
+            filtersData: view.layout.filters,
+            filtersActive: response.data.filters,
+          });
+
+          dispatch(
+            createFilter({
+              filterId,
+              data: {
+                filterData: view.layout.filters, // set the proper layout for the filters
+                filtersActive,
+                activeFiltersCaptions,
+              },
+            })
+          );
+        }
 
         let shouldFetchQuickActions = true;
         let viewProfileId = null;
@@ -416,7 +423,7 @@ export function fetchDocument({
         }
 
         // get quickactions
-        if (viewId && shouldFetchQuickActions) {
+        if (viewId && shouldFetchQuickActions && !websocketRefresh) {
           dispatch(
             fetchQuickActions({
               windowId,
